upgrade to spring-boot 2.5.8 and telegrambots 6.0.1 - change data types of some telegram API entity fields - disable telegram by default (as it requires real bot token/username) - add empty default values for telegram properties - fix test assertions to JUnit5 - add re-routing messages to user in test_local profile (see *Spy)

+++ src/main/java/org/vors/pairbot/service/UserService.kt
@@ -37,19 +37,19 @@ open class UserService(
         return createUserInfo(userId, firstName, lastName)
     }
 
-    private fun createUserInfo(userId: Int, firstName: String, lastName: String): UserInfo {
+    private fun createUserInfo(userId: Long, firstName: String, lastName: String): UserInfo {
         return newUserInfo(userId, firstName, lastName)
     }
 
-    private fun newUserInfo(id: Int, firstName: String, lastName: String): UserInfo {
+    private fun newUserInfo(id: Long, firstName: String, lastName: String): UserInfo {
         return UserInfo(id, firstName, lastName, createdDate = Date())
     }
 
-    fun getExistingUser(userId: Int?): UserInfo {
+    fun getExistingUser(userId: Long?): UserInfo {
         return findByUserId(userId).orElseThrow { IllegalStateException("Existing user not found for ID: " + userId!!) }
     }
 
-    fun findByUserId(userId: Int?): Optional<UserInfo> {
+    fun findByUserId(userId: Long?): Optional<UserInfo> {
         return userRepository.findByUserId(userId)
     }
